Initial commit for Spark batch ingestion common code abstraction.
