+++ pinot-plugins/pinot-batch-ingestion/pinot-batch-ingestion-spark-base/pom.xml
@@ -27,8 +27,8 @@
     <version>1.3.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>pinot-batch-ingestion-spark-3</artifactId>
-  <name>Pinot Batch Ingestion for Spark 3</name>
+  <artifactId>pinot-batch-ingestion-spark-base</artifactId>
+  <name>Pinot Batch Ingestion Spark Base</name>
   <url>https://pinot.apache.org/</url>
   <properties>
     <pinot.root>${basedir}/../../..</pinot.root>
@@ -39,44 +39,6 @@
     <dependency>
       <groupId>org.apache.pinot</groupId>
       <artifactId>pinot-batch-ingestion-common</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.spark</groupId>
-      <artifactId>spark-core_${scala.compat.version}</artifactId>
-      <version>${spark3.version}</version>
-      <scope>provided</scope>
-      <exclusions>
-        <exclusion>
-          <groupId>commons-logging</groupId>
-          <artifactId>commons-logging</artifactId>
-        </exclusion>
-        <exclusion>
-          <groupId>log4j</groupId>
-          <artifactId>log4j</artifactId>
-        </exclusion>
-        <exclusion>
-          <groupId>org.slf4j</groupId>
-          <artifactId>slf4j-log4j12</artifactId>
-        </exclusion>
-      </exclusions>
-    </dependency>
-    <dependency>
-      <groupId>org.scala-lang</groupId>
-      <artifactId>scala-library</artifactId>
-      <scope>provided</scope>
-    </dependency>
-
-    <!-- For testing import of csv files -->
-    <dependency>
-      <groupId>org.apache.pinot</groupId>
-      <artifactId>pinot-csv</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.pinot</groupId>
-      <artifactId>pinot-batch-ingestion-spark-base</artifactId>
-      <version>1.3.0-SNAPSHOT</version>
-      <scope>compile</scope>
     </dependency>
   </dependencies>
