+++ pinot-plugins/pinot-batch-ingestion/pom.xml
@@ -37,6 +37,7 @@
 
   <modules>
     <module>pinot-batch-ingestion-common</module>
+    <module>pinot-batch-ingestion-spark-base</module>
     <module>pinot-batch-ingestion-spark-2.4</module>
     <module>pinot-batch-ingestion-spark-3</module>
